Add top border and disable caching
Caching has been disabled for developing purposes. This is done by
setting the appropriate header files in the index.html file and by
adding an artificial query parameter to the javascript file, whose only
purpose is to let the browser believe a new file is requested.

From a functional point of view, a top border has been added that is
loadd from a png file.